requirements.txt
anthropic>=0.25.0
langgraph>=0.1.0
langchain-anthropic>=0.1.0
openai-whisper>=20231117
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.0.0
